+++ libraries/decoder/src/main/java/androidx/media3/decoder/DecoderOutputBuffer.java
@@ -15,6 +15,7 @@
  */
 package androidx.media3.decoder;
 
+import androidx.annotation.CallSuper;
 import androidx.media3.common.util.UnstableApi;
 
 /** Output buffer decoded by a {@link Decoder}. */
@@ -40,6 +41,21 @@ public abstract class DecoderOutputBuffer extends Buffer {
    */
   public int skippedOutputBufferCount;
 
+  /**
+   * Whether this buffer should be skipped, usually because the decoding process generated no data
+   * or invalid data.
+   */
+  public boolean shouldBeSkipped;
+
   /** Releases the output buffer for reuse. Must be called when the buffer is no longer needed. */
   public abstract void release();
+
+  @Override
+  @CallSuper
+  public void clear() {
+    super.clear();
+    timeUs = 0;
+    skippedOutputBufferCount = 0;
+    shouldBeSkipped = false;
+  }
 }