+++ libraries/decoder/src/main/java/androidx/media3/decoder/SimpleDecoder.java
@@ -232,6 +232,7 @@ public abstract class SimpleDecoder<
     if (inputBuffer.isEndOfStream()) {
       outputBuffer.addFlag(C.BUFFER_FLAG_END_OF_STREAM);
     } else {
+      outputBuffer.timeUs = inputBuffer.timeUs;
       if (inputBuffer.isDecodeOnly()) {
         outputBuffer.addFlag(C.BUFFER_FLAG_DECODE_ONLY);
       }
@@ -262,7 +263,7 @@ public abstract class SimpleDecoder<
     synchronized (lock) {
       if (flushed) {
         outputBuffer.release();
-      } else if (outputBuffer.isDecodeOnly()) {
+      } else if (outputBuffer.isDecodeOnly() || outputBuffer.shouldBeSkipped) {
         skippedOutputBufferCount++;
         outputBuffer.release();
       } else {