+++ libraries/decoder_opus/src/main/java/androidx/media3/decoder/opus/OpusDecoder.java
@@ -229,7 +229,7 @@ public final class OpusDecoder
       int skipBytes = skipSamples * bytesPerSample;
       if (result <= skipBytes) {
         skipSamples -= result / bytesPerSample;
-        outputBuffer.addFlag(C.BUFFER_FLAG_DECODE_ONLY);
+        outputBuffer.shouldBeSkipped = true;
         outputData.position(result);
       } else {
         skipSamples = 0;